requirements.txt
altair
attrs
blinker
blis
cachetools
catalogue
certifi
cffi
chardet
charset-normalizer
click
colorama
commonmark
cryptography
cycler
cymem
decorator
docx2txt
entrypoints
fonttools
future
geocoder
geographiclib
geopy
gitdb
GitPython
idna
importlib-metadata
Jinja2
joblib
jsonschema
kiwisolver
langcodes
MarkupSafe
matplotlib
murmurhash
nltk
numpy
packaging
pafy
pandas
pathy
pdfminer.six
pdfminer3
Pillow
plac
plotly
preshed
protobuf
pyarrow
pycparser
pycryptodome
pydantic
pydeck
Pygments
Pympler
PyMySQL
pyparsing
pyresparser
pyrsistent
python-dateutil
pytz
pytz-deprecation-shim
ratelim
regex
requests
rich
semver
six
smart-open
smmap
sortedcontainers
spacy
spacy-legacy
spacy-loggers
srsly
streamlit
streamlit-tags
tenacity
thinc
toml
toolz
tornado
tqdm
typer
typing_extensions
tzdata
tzlocal
urllib3
validators
wasabi
watchdog
zipp
